% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data4}
\alias{data4}
\title{Simulated Dataset 4}
\format{
A data frame with 190 rows and 2 variables:
\describe{
\item{x1}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\source{
Simulated data generated with the script provided in the inst directory.
}
\usage{
data4
}
\description{
Unbounded response, no rshp between response and predictor, covariance in residuals
}
\keyword{datasets}
