% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_stan.R
\name{linreg_stan}
\alias{linreg_stan}
\title{linreg_stan}
\usage{
linreg_stan(
  des,
  y,
  model = "ols",
  covmat = NULL,
  iter = 2000,
  chains = 4,
  cores = 4,
  coef.u = 0,
  coef.sd = 10,
  physig.u = -1,
  physig.sd = 1,
  randsig.loc = 0,
  randsig.sc = 2.5,
  ...
)
}
\arguments{
\item{des}{A vector of predictor variable observations OR, in the case of multiple predictors, a matrix in which each column is a vector of observations of a given predictor. 'betareg.stan' adds a column of 1s to make this a design matrix whose first column corresponds to the model intercept (unless such a column already exists).}

\item{y}{A vector of response variable observations.}

\item{model}{Choice of "ols", "pgls", or "pgls.mm"; defaults to ols.}

\item{iter}{Number of iterations to run on each chain; defaults to 2000 (more are often necessary).}

\item{chains}{Number of MCMC chains to run; defaults to 4.}

\item{cores}{Number of cores to use.}

\item{coef.u}{Mean of the Gaussian prior for each preditor variable coefficient; defaults to 0.}

\item{coef.sd}{SD of the Gaussian prior for each preditor variable coefficient; defaults to 10.}

\item{physig.u}{Mean of the prior distribution (lognormal) for the scale of the phylogenetic component of residual covariance; defaults to -1.}

\item{physig.sd}{SD of the prior distribution (lognormal) for the scale of the phylogenetic component of residual covariance; defaults to 1.}

\item{randsig.loc}{Location parameter of the prior distribution (cauchy) for the scale of the independent component of residual covariance ; defaults to 0.}

\item{randsig.sc}{Scale parameter of the prior distribution (cauchy) for the scale of the independent component of residual covariance ; defaults to 2.5.}

\item{...}{additional arguments passed to \code{rstan::sampling}, including control parmeters (see rstan::sampling documentation)}

\item{cov}{Covariance matrix for model residuals (a lineage-pair covariance matrix if analyzing lineage-pair data or a phylogenetic vcv matrix if analyzing species data).}
}
\value{
A list containing two elements: (1) the posterior distribution of beta model parameters, and (2) the log-likelihood of the posteriors for use in downstream analyses (e.g. the calculation of model fitting metrics like loo or waic)
}
\description{
Fits one of three models of linear regression to a dataset in a Bayesian
framework. Bayesian sampling is conducted in the Stan software via the 'rstan'
package. Users supply vectors of observations and, unless an "ols" model is chosen,
a covariance matrix. Users can alter parameters for model-parameter prior distributions
and MCMC sampling settings. See details.
}
\details{
The models that can be chosen are:
\enumerate{
\item "ols": basic ordinary least-squares regression
\code{Y = XB + eps; eps~N(0,randsig*I)}  (randsig is sigma^2 scaling parameter for identity matrix I)
\item "pgls": phylogenetic generalized least-squares regression in which residuals
covary according to covariance matrix C. This is a lineage-pair covariance
matrix (if analyzing lineage-pair data) or a phylogenetic vcv matrix
(if analyzing species data).
\code{Y = XB + eps; eps~N(0,physig*C)} (physig is the sigma^2 scaling parameter for covariance matrix C)
\item "pgls.mm" a mixed-effects version of pgls in which the residuals have
uncorrelated and correlated components, where the latter are structured according
to covariance matrix C.
\code{Y = XB + eps + u; eps~N(0,randsig*I), u~N(0,physig*C)}
}
}
\section{NOTE: For models 2 and 3, the user must supply a covariance matrix.}{

}

\section{Prior Distributions for Regression Model Parameters}{
The underlying stan models assume the following prior distributions for model parameters
\enumerate{
\item Regression coefficients: Gaussian prior (users can set prior mean and sd)
\item \code{phylosig}: lognormal prior (users can set prior mean and sd)
\item \code{randsig}: cauchy prior (users can set location and scale parameters of prior)
}
}

\examples{
## Fit regression models with and without covariance matrix
# Note: data were simulated with Coef[1]=1 (intercept), Coef[2]=0.8 (slope)
# Load a data simulated with a non-independent response observations
data(data3)
# Also load the lineage-pair covariance matrix that arose from those simulations
data(sim.cov.pairs)
# Fit an OLS model
result1 = linreg_stan(des=data3[,1], y=data3[,2])
# Fit an pgls model
result2 = linreg_stan(des=data3[,1], y=data3[,2], model="pgls", covmat=sim.cov.pairs)
# Compare posterior parameter estimates
result1[[1]]
result2[[1]]
# Compare the fit of the two models via loo and waic
loo1 = loo::loo(result1[[2]])
loo2 = loo::loo(result2[[2]])
waic1 = loo::waic(result1[[2]])
waic2 = loo::waic(result2[[2]])
loo1
loo2
waic1
waic2
loo::loo_compare(loo1, loo2)
loo::loo_compare(waic1, waic2)
\dontrun{
# Extend the comparison by fitting a pgls.mm model
result3 = linreg_stan(des=data3[,1], y=data3[,2], model="pgls.mm", covmat=sim.cov.pairs)
# Compare posterior parameter estimates
result1[[1]]
result2[[1]]
result3[[1]]
# Compare the fit of the three models via loo and waic
loo3 = loo::loo(result3[[2]])
waic3 = loo::waic(result3[[2]])
loo::loo_compare(loo1, loo2, loo3)
loo::loo_compare(waic1, waic2, waic3)
}
}
