% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.cov.pairs}
\alias{sim.cov.pairs}
\title{Simulated lineage-pair covariance matrix}
\format{
A matrix of 190 rows and 190 columns
}
\source{
Simulated data generated with the script provided in the inst directory.
}
\usage{
sim.cov.pairs
}
\description{
A 190*190 covariance matrix used in simulating the example datasets in this package.
}
\keyword{datasets}
