% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated.datasets}
\alias{simulated.datasets}
\alias{data1}
\alias{data2}
\alias{data3}
\alias{data4}
\alias{data5}
\alias{data6}
\alias{data7}
\alias{data8}
\alias{sim.cov.pairs}
\alias{sim.tree1}
\title{Simulated Datasets}
\format{
The datasets are as follows
\describe{
\item{data1}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data2}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data3}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data4}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data5}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data6}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data7}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{data8}{A data frame with 190 rows and 4 columns.
\describe{
\item{sp1}{Identity of Species 1 in Pair}
\item{sp2}{Identity of Species 2 in Pair}
\item{pred}{Numeric predictor variable}
\item{y}{Numeric response variable}
}
}
\item{sim.cov.pairs}{A 190x190 covariance matrix
}
\item{sim.tree1}{A phylo object
}
}
}
\source{
Simulated data generated with the script provided in the 'inst' directory.
}
\description{
Simulated datasets are provided to help users understand functions. Lineage-pairs (n=190) were created from a simulated phylogenetic tree with 20 tips. A predictor variable was generated from random draws of a standard normal. Reponse variables of various structures were simulated as follows:
\enumerate{
\item \strong{Simulated Dataset 1} Unbounded response, linear relationship between response and predictor, no covariance in residuals
\item \strong{Simulated Dataset 2} Unbounded response, no relationship between response and predictor, no covariance in residuals
\item \strong{Simulated Dataset 3} Unbounded response, linear relationship between response and predictor, covariance in residuals
\item \strong{Simulated Dataset 4} Unbounded response, no relationship between response and predictor, covariance in residuals
\item \strong{Simulated Dataset 5} Bounded response, linear relationship between response and predictor on link scale, no covariance in residuals
\item \strong{Simulated Dataset 6} Bounded response, no relationship between response and predictor on link scale, no covariance in residuals
\item \strong{Simulated Dataset 7} Bounded response, linear relationship between response and predictor on link scale, covariance in residuals
\item \strong{Simulated Dataset 8} Bounded response, no relationship between response and predictor on link scale, covariance in residuals
\item \strong{Simulated Lineage-Pair Covariance Matrix} A 190*190 covariance matrix used in simulating the example datasets in this package.
\item \strong{Simulated Phylogenetic Tree} A 20-species ultrametric phylogenetic tree used in simulating the example datasets in this package.
}
}
