% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopairs-package.R
\docType{_PACKAGE}
\name{phylopairs-package}
\alias{phylopairs-package}
\alias{phylopairs}
\title{The 'phylopairs' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
